pyproject.toml
[tool.poetry]
name = "bananaproto"
version = "0.0.0"
description = "A BANANA Protobuf / gRPC generator & library"
authors = ["Daniel G. Taylor <danielgtaylor@gmail.com>"]
maintainers = ["BananaLoaf <bananaloaf@protonmail.com>"]
readme = "README.md"
repository = "https://github.com/BananaLoaf/python-bananaproto"
keywords = ["protobuf", "gRPC"]
license = "MIT"
packages = [
    { include = "bananaproto", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
black = { version = ">=19.3b0", optional = true }
grpclib = "^0.4.1"
jinja2 = { version = ">=3.0.3", optional = true }
python-dateutil = "^2.8"
isort = {version = "^5.11.5", optional = true}
python-singleton = "^0.1.2"

[tool.poetry.dev-dependencies]
bpython = "^0.19"
grpcio-tools = "^1.50.0"
jinja2 = ">=3.0.3"
mypy = "^0.930"
poethepoet = ">=0.9.0"
protobuf = "^4.21.6"
pytest = "^6.2.5"
pytest-asyncio = "^0.12.0"
pytest-cov = "^2.9.0"
pytest-mock = "^3.1.1"
tomlkit = "^0.7.0"
black = ">=19.3b0"
isort = "^5.11.5"


[tool.poetry.scripts]
protoc-gen-python_bananaproto = "bananaproto.plugin:main"

[tool.poetry.extras]
compiler = ["black", "isort", "jinja2"]


# Dev workflow tasks

[tool.poe.tasks.generate]
script = "tests.generate:main"
help = "Generate test cases (do this once before running test)"

[tool.poe.tasks.types]
cmd  = "mypy src --ignore-missing-imports"
help = "Check types with mypy"

[tool.poe.tasks.format]
cmd  = "black . --exclude output_"
help = "Apply black formatting to source code"

[tool.poe.tasks.check-style]
cmd = "black . --check --diff"
help = "Check if code style is correct"

[tool.poe.tasks.clean]
cmd  = """
rm -rf .asv .coverage .mypy_cache .pytest_cache
       dist bananaproto.egg-info **/__pycache__
       testsoutput_* **/output_bananaproto
       **/output_reference
"""
help = "Clean out generated files from the workspace"

[tool.poe.tasks.generate_lib]
cmd = """
protoc
    --plugin=protoc-gen-custom=src/bananaproto/plugin/main.py
    --custom_opt=INCLUDE_GOOGLE
    --custom_out=src/bananaproto/lib
    -I /usr/local/include/
    /usr/local/include/google/protobuf/**/*.proto
"""
help = "Regenerate the types in bananaproto.lib.google"


[tool.isort]
py_version = 37
profile = "black"
force_single_line = false
combine_as_imports = true
lines_after_imports = 2
include_trailing_comma = true
force_grid_wrap = 2
src_paths = ["src", "tests"]

[tool.black]
target-version = ['py37']

[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = ""
norecursedirs = "**/output_*"
addopts = "-p no:warnings"

[build-system]
requires = ["poetry-core>=1.0.0,<2"]
build-backend = "poetry.core.masonry.api"
